C3/C4: ring (Flash-style) attention, OpenMP threading, collective hybrid; finish C2
Algorithm & parallelism upgrades:
- C3 ring/Flash tensor path (--ring): K,V sharded O(N/P) per rank and streamed
  around a ring, folded into a running online softmax. Selected via a new
  TPVariant enum (replaces the use_cannon int). Correct ~1e-7 at any P.
- C5 non-blocking overlap (folded into ring): double-buffered MPI_Isend/Irecv +
  Waitall so each K,V shift overlaps the per-block compute.
- C4 OpenMP: -fopenmp + parallel-for over query rows (per-thread scratch) ->
  hybrid MPI+OpenMP; prints threads=. Kernel is memory-bandwidth bound.
- Hybrid: replace the serial rank-0 Send/Recv fan-out with MPI_Scatter/Gather
  over a group-roots sub-communicator (O(log G) vs O(G)). Still bitwise-exact.

Finish C2 profiler wiring (was in progress):
- Parse --profile-wait, call the latency probe, fix the CSV header to 9 columns.
- Bench scripts emit the 9-col schema; granularity uses --profile-wait; speedup
  captures msgs/bytes/latency. make_charts surfaces wait/packets/bytes/latency,
  plus new ring-vs-1D and OpenMP thread-scaling charts. N sweep extended to 65536.

Hardening & cleanup:
- data_gen: size_t to avoid int overflow at large N.
- Remove dead tensor helpers (matmul/softmax_rows/tensor_scale/add/copy).
- README + REPORT.tex updated for ring, OpenMP, and the collective hybrid.

New scripts: scripts/bench_ring_vs_1d.sh, scripts/bench_threads.sh

+++ src/data_gen.c
@@ -3,7 +3,7 @@
 
 void data_gen_fill(Tensor *t, unsigned int seed) {
     srand(seed);
-    int n = t->rows * t->cols;
-    for (int i = 0; i < n; i++)
+    size_t n = (size_t)t->rows * t->cols;   /* size_t: avoid int overflow at large N */
+    for (size_t i = 0; i < n; i++)
         t->data[i] = ((float)rand() / (float)RAND_MAX) * 2.0f - 1.0f;
 }
